requirements.txt
discord.py
gspread
oauth2client
aioschedule
matplotlib